refactor: cleanup gameplay logic
-Add logic to display list of words entered by user when they toggle the capslock key or press and hold the spacebar key.

-Create relevant hooks to manage existing gameplay logic for better reusability.
